+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+* feature: support for regions endpoint
 * breaking change: `Environment.bulkDestroy` returns an array of ids instead of an object (with an array of  ids)
 * feature: list processes, run and stop one-off containers
 * Improved type definitions